[broken] Move sdk into own crate - need to fix dev deps issue
