+++ src/core_tests/replay_flag.rs
@@ -1,10 +1,7 @@
-use crate::{
-    prototype_rust_sdk::{WfContext, WorkflowFunction},
-    test_help::canned_histories,
-    workflow::managed_wf::ManagedWFFunc,
-};
+use crate::{test_help::canned_histories, workflow::managed_wf::ManagedWFFunc};
 use rstest::{fixture, rstest};
 use std::time::Duration;
+use temporal_sdk::{WfContext, WorkflowFunction};
 use temporal_sdk_core_protos::temporal::api::enums::v1::CommandType;
 
 fn timers_wf(num_timers: u32) -> WorkflowFunction {
