+++ src/workflow/machines/cancel_external_state_machine.rs
@@ -230,11 +230,8 @@ impl Cancellable for CancelExternalMachine {}
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        prototype_rust_sdk::{WfContext, WorkflowFunction, WorkflowResult},
-        test_help::TestHistoryBuilder,
-        workflow::managed_wf::ManagedWFFunc,
-    };
+    use crate::{test_help::TestHistoryBuilder, workflow::managed_wf::ManagedWFFunc};
+    use temporal_sdk::{WfContext, WorkflowFunction, WorkflowResult};
 
     async fn cancel_sender(ctx: WfContext) -> WorkflowResult<()> {
         let res = ctx
