+++ src/workflow/machines/signal_external_state_machine.rs
@@ -285,12 +285,9 @@ impl Cancellable for SignalExternalMachine {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        prototype_rust_sdk::{CancellableFuture, WfContext, WorkflowFunction, WorkflowResult},
-        test_help::TestHistoryBuilder,
-        workflow::managed_wf::ManagedWFFunc,
-    };
+    use crate::{test_help::TestHistoryBuilder, workflow::managed_wf::ManagedWFFunc};
     use std::mem::discriminant;
+    use temporal_sdk::{CancellableFuture, WfContext, WorkflowFunction, WorkflowResult};
     use temporal_sdk_core_protos::coresdk::workflow_activation::{
         wf_activation_job, WfActivationJob,
     };
