+++ temporal-sdk/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "temporal-sdk"
+version = "0.1.0-alpha.1"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+crossbeam = "0.8"
+derive_more = "0.99"
+futures = "0.3"
+parking_lot = { version = "0.11", features = ["send_guard"] }
+prost-types = "0.9"
+serde = "1.0"
+tokio = { version = "1.1", features = ["rt", "rt-multi-thread", "parking_lot", "time", "fs"] }
+tokio-util = { version = "0.6.9" }
+tokio-stream = "0.1"
+tonic = "0.6"
+tracing = { version = "0.1", features = ["log-always"] }
+
+[dependencies.temporal-sdk-core]
+path = ".."
+version = "0.1"
+
+[dependencies.temporal-sdk-core-protos]
+path = "../sdk-core-protos"
+version = "0.1"
