+++ test_utils/Cargo.toml
@@ -14,6 +14,7 @@ futures = "0.3"
 log = "0.4"
 rand = "0.8"
 temporal-sdk-core = { path = ".." }
+temporal-sdk = { path = "../temporal-sdk" }
 url = "2.2"
 
 [dependencies.temporal-sdk-core-protos]
