+++ tests/integ_tests/workflow_tests/cancel_wf.rs
@@ -1,5 +1,5 @@
 use std::time::Duration;
-use temporal_sdk_core::prototype_rust_sdk::{WfContext, WfExitValue, WorkflowResult};
+use temporal_sdk::{WfContext, WfExitValue, WorkflowResult};
 use temporal_sdk_core_protos::temporal::api::enums::v1::WorkflowExecutionStatus;
 use test_utils::CoreWfStarter;
