+++ tests/integ_tests/workflow_tests/continue_as_new.rs
@@ -1,5 +1,5 @@
 use std::time::Duration;
-use temporal_sdk_core::prototype_rust_sdk::{WfContext, WfExitValue, WorkflowResult};
+use temporal_sdk::{WfContext, WfExitValue, WorkflowResult};
 use temporal_sdk_core_protos::coresdk::workflow_commands::ContinueAsNewWorkflowExecution;
 use test_utils::CoreWfStarter;
